servo: Update README and improve code (#11447)
* refactor: use wgpu re-exported from slint and remove redundant wgpu-hal/wgpu dependencies

* refactor: improve lib

* docs: Update README for new implementation

* refactor: move get_wgpu_texture_from_metal from GPURenderingContext to metal module extension trait

* refactor: fix wgpu_hal import to wpgu::hal from slint

* hal

* refactor: decouple android and linux gpu rendering context configurations to support platform-specific API detection

* [autofix.ci] apply automated fixes

---------
