+++ examples/servo/src/webview/rendering_context/gpu_rendering_context.rs
@@ -5,6 +5,7 @@ use std::{cell::Cell, rc::Rc, sync::Arc};
 
 use euclid::default::Size2D;
 use image::RgbaImage;
+use slint::wgpu_28::wgpu;
 use winit::dpi::PhysicalSize;
 
 use servo::{DeviceIntRect, RenderingContext};
@@ -149,7 +150,7 @@ impl GPURenderingContext {
                     "Unknown"
                 }
             }
-            #[cfg(any(target_os = "linux", target_os = "android"))]
+            #[cfg(target_os = "linux")]
             {
                 use api::{Gles, Vulkan};
                 if wgpu_device.as_hal::<Vulkan>().is_some() {
@@ -160,6 +161,11 @@ impl GPURenderingContext {
                     "Unknown"
                 }
             }
+            #[cfg(target_os = "android")]
+            {
+                use api::Vulkan;
+                if wgpu_device.as_hal::<Vulkan>().is_some() { "Vulkan" } else { "Unknown" }
+            }
             #[cfg(target_vendor = "apple")]
             {
                 use api::Metal;
@@ -167,39 +173,6 @@ impl GPURenderingContext {
             }
         };
         eprintln!("[GPU] Active WGPU backend: {}", backend);
-    }
-
-    /// Imports Metal surface as a WGPU texture for rendering on macOS/iOS.
-    /// Unbinds the surface, converts to WGPU texture, then rebinds it.
-    #[cfg(target_vendor = "apple")]
-    pub fn get_wgpu_texture_from_metal(
-        &self,
-        wgpu_device: &wgpu::Device,
-        wgpu_queue: &wgpu::Queue,
-    ) -> Result<wgpu::Texture, surfman::Error> {
-        use super::metal::WPGPUTextureFromMetal;
-
-        let device = &self.surfman_rendering_info.device.borrow();
-        let mut context = self.surfman_rendering_info.context.borrow_mut();
-
-        let surface = device.unbind_surface_from_context(&mut context)?.unwrap();
-
-        let size = self.size.get();
-
-        let wgpu_texture = WPGPUTextureFromMetal::new(size, wgpu_device).get(
-            wgpu_device,
-            wgpu_queue,
-            device,
-            &surface,
-        );
-
-        let _ =
-            device.bind_surface_to_context(&mut context, surface).map_err(|(err, mut surface)| {
-                let _ = device.destroy_surface(&mut context, &mut surface);
-                err
-            });
-
-        Ok(wgpu_texture)
     }
 }
