+++ examples/servo/src/webview/rendering_context/metal/mod.rs
@@ -4,5 +4,4 @@
 mod metal;
 mod texture_importer;
 
-pub use metal::WPGPUTextureFromMetal;
 pub use texture_importer::ServoTextureImporter;
